Fixed bug not allowing editing of artist name or title.
