+++ package.json
@@ -54,6 +54,6 @@
 		"typescript": "^4.7.4",
 		"vite": "^2.9.12",
 		"vue-eslint-parser": "^9.0.2",
-		"vue-tsc": "^0.38.0"
+		"vue-tsc": "^0.38.1"
 	}
 }
